Add first prototype implementation for guided forms and form elements
This suffers currently from not being able to populate the results of a guide
to its form or form element. Additionally it is not that flexible to be actually
utilized due to some tripping stones in the wizard implementation. We
really need to think about how to implement "mini-wizards" properly and
how they are going to be utilized...

refs #9603